Implemented per service mapping
Fixes #24

This implements per service mapping of service name, port, protocol and
scheme.

The `ServiceNameParser` has been renamed to `ServiceNameMapper`, because
that's now what it essentially does, maps a Lagom service name to an
Akka discover lookup, plus scheme. It first attempts to do the mapping
via direct configuration. If there's no mapping there, it then attempts
to do the mapping parsing it using the regex, and if that yields no
mapping, then it just uses the service name as is with the configured
default port, protocol and mapped scheme.